adjust project handlers
