+++ services/Cargo.toml
@@ -64,6 +64,7 @@ proj = "0.22"
 proj-sys = "0.19"
 pwhash = "1.0"
 quick-xml = { version = "0.28", optional = true }
+rand = "0.8"
 rayon = "1.5"
 regex = "1.5"
 reqwest = { version = "0.11", features = ["json", "multipart", "stream"] }
@@ -100,7 +101,6 @@ assert_cmd = "2.0"
 httptest = "0.15.2"
 hyper = "0.14"
 prost = "0.11"
-rand = "0.8.4"
 serial_test = "2.0"
 tempfile = "3.1"
 xml-rs = "0.8"
