+++ services/src/contexts/postgres.rs
@@ -624,13 +624,12 @@ mod tests {
         ProjectDb, ProjectFilter, ProjectId, ProjectLayer, ProjectListOptions, ProjectListing,
         STRectangle, UpdateProject,
     };
-    use crate::util::config::{get_config_element, Postgres};
     use crate::util::tests::register_ndvi_workflow_helper;
+    use crate::util::tests::with_temp_context;
     use crate::workflows::registry::WorkflowRegistry;
     use crate::workflows::workflow::Workflow;
-    use bb8_postgres::bb8::ManageConnection;
-    use bb8_postgres::tokio_postgres::{self, NoTls};
-    use futures::{join, Future};
+    use bb8_postgres::tokio_postgres::NoTls;
+    use futures::join;
     use geoengine_datatypes::collections::VectorDataType;
     use geoengine_datatypes::primitives::CacheTtlSeconds;
     use geoengine_datatypes::primitives::{
@@ -640,7 +639,6 @@ mod tests {
     };
     use geoengine_datatypes::raster::RasterDataType;
     use geoengine_datatypes::spatial_reference::{SpatialReference, SpatialReferenceOption};
-    use geoengine_datatypes::util::test::TestDefault;
     use geoengine_operators::engine::{
         MetaData, MetaDataProvider, MultipleRasterOrSingleVectorSource, PlotOperator,
         RasterResultDescriptor, StaticMetaData, TypedOperator, TypedResultDescriptor,
@@ -655,86 +653,7 @@ mod tests {
         OgrSourceDatasetTimeType, OgrSourceDurationSpec, OgrSourceErrorSpec, OgrSourceTimeFormat,
     };
     use geoengine_operators::util::input::MultiRasterOrVectorOperator::Raster;
-    use rand::RngCore;
     use serde_json::json;
-    use tokio::runtime::Handle;
-
-    /// Setup database schema and return its name.
-    async fn setup_db() -> (tokio_postgres::Config, String) {
-        let mut db_config = get_config_element::<Postgres>().unwrap();
-        db_config.schema = format!("geoengine_test_{}", rand::thread_rng().next_u64()); // generate random temp schema
-
-        let mut pg_config = tokio_postgres::Config::new();
-        pg_config
-            .user(&db_config.user)
-            .password(&db_config.password)
-            .host(&db_config.host)
-            .dbname(&db_config.database);
-
-        // generate schema with prior connection
-        PostgresConnectionManager::new(pg_config.clone(), NoTls)
-            .connect()
-            .await
-            .unwrap()
-            .batch_execute(&format!("CREATE SCHEMA {};", &db_config.schema))
-            .await
-            .unwrap();
-
-        // fix schema by providing `search_path` option
-        pg_config.options(&format!("-c search_path={}", db_config.schema));
-
-        (pg_config, db_config.schema)
-    }
-
-    /// Tear down database schema.
-    async fn tear_down_db(pg_config: tokio_postgres::Config, schema: &str) {
-        // generate schema with prior connection
-        PostgresConnectionManager::new(pg_config, NoTls)
-            .connect()
-            .await
-            .unwrap()
-            .batch_execute(&format!("DROP SCHEMA {schema} CASCADE;"))
-            .await
-            .unwrap();
-    }
-
-    async fn with_temp_context<F, Fut>(f: F)
-    where
-        F: FnOnce(PostgresContext<NoTls>, tokio_postgres::Config) -> Fut
-            + std::panic::UnwindSafe
-            + Send
-            + 'static,
-        Fut: Future<Output = ()> + Send,
-    {
-        let (pg_config, schema) = setup_db().await;
-
-        // catch all panics and clean up first…
-        let executed_fn = {
-            let pg_config = pg_config.clone();
-            std::panic::catch_unwind(move || {
-                tokio::task::block_in_place(move || {
-                    Handle::current().block_on(async move {
-                        let ctx = PostgresContext::new_with_context_spec(
-                            pg_config.clone(),
-                            tokio_postgres::NoTls,
-                            TestDefault::test_default(),
-                            TestDefault::test_default(),
-                        )
-                        .await
-                        .unwrap();
-                        f(ctx, pg_config.clone()).await;
-                    });
-                });
-            })
-        };
-
-        tear_down_db(pg_config, &schema).await;
-
-        // then throw errors afterwards
-        if let Err(err) = executed_fn {
-            std::panic::resume_unwind(err);
-        }
-    }
 
     #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
     async fn test() {
