+++ services/src/handlers/projects.rs
@@ -1,7 +1,10 @@
 use crate::contexts::ApplicationContext;
 use crate::error::Result;
 use crate::handlers::SessionContext;
-use crate::projects::{CreateProject, ProjectDb, ProjectId, ProjectListOptions, UpdateProject};
+use crate::projects::{
+    CreateProject, LoadVersion, ProjectDb, ProjectId, ProjectListOptions, ProjectVersionId,
+    UpdateProject,
+};
 use crate::util::extractors::{ValidatedJson, ValidatedQuery};
 use crate::util::IdResponse;
 use actix_web::{web, FromRequest, HttpResponse, Responder};
@@ -11,14 +14,24 @@ where
     C: ApplicationContext,
     C::Session: FromRequest,
 {
-    // TODO: add project versions now that the free backend supports it
-    cfg.service(web::resource("/project").route(web::post().to(create_project_handler::<C>)))
-        .service(web::resource("/projects").route(web::get().to(list_projects_handler::<C>)))
+    cfg.service(web::resource("/projects").route(web::get().to(list_projects_handler::<C>)))
         .service(
-            web::resource("/project/{project}")
-                .route(web::get().to(load_project_handler::<C>))
-                .route(web::patch().to(update_project_handler::<C>))
-                .route(web::delete().to(delete_project_handler::<C>)),
+            web::scope("/project")
+                .service(web::resource("").route(web::post().to(create_project_handler::<C>)))
+                .service(
+                    web::resource("/{project}")
+                        .route(web::get().to(load_project_latest_handler::<C>))
+                        .route(web::patch().to(update_project_handler::<C>))
+                        .route(web::delete().to(delete_project_handler::<C>)),
+                )
+                .service(
+                    web::resource("/{project}/versions")
+                        .route(web::get().to(project_versions_handler::<C>)),
+                )
+                .service(
+                    web::resource("/{project}/{version}")
+                        .route(web::get().to(load_project_version_handler::<C>)),
+                ),
         );
 }
 
@@ -87,68 +100,6 @@ pub(crate) async fn list_projects_handler<C: ApplicationContext>(
     Ok(web::Json(listing))
 }
 
-/// Retrieves details about the latest version of a project.
-#[utoipa::path(
-    tag = "Projects",
-    get,
-    path = "/project/{project}",
-    responses(
-        (status = 200, description = "Project loaded from database", body = Project,
-            example = json!({
-                "id": "df4ad02e-0d61-4e29-90eb-dc1259c1f5b9",
-                "version": {
-                    "id": "8f4b8683-f92c-4129-a16f-818aeeee484e",
-                    "changed": "2021-04-26T14:05:39.677390600Z",
-                    "author": "5b4466d2-8bab-4ed8-a182-722af3c80958"
-                },
-                "name": "Test",
-                "description": "Foo",
-                "layers": [],
-                "plots": [],
-                "bounds": {
-                    "spatialReference": "EPSG:4326",
-                    "boundingBox": {
-                        "lowerLeftCoordinate": {
-                            "x": 0.0,
-                            "y": 0.0
-                        },
-                        "upperRightCoordinate": {
-                            "x": 1.0,
-                            "y": 1.0
-                        }
-                    },
-                    "timeInterval": {
-                        "start": 0,
-                        "end": 1
-                    }
-                },
-                "timeStep": {
-                    "granularity": "months",
-                    "step": 1
-                }
-            })
-        )
-    ),
-    params(
-        ("project" = ProjectId, description = "Project id")
-    ),
-    security(
-        ("session_token" = [])
-    )
-)]
-async fn load_project_handler<C: ApplicationContext>(
-    project: web::Path<ProjectId>,
-    session: C::Session,
-    app_ctx: web::Data<C>,
-) -> Result<impl Responder> {
-    let id = app_ctx
-        .session_context(session)
-        .db()
-        .load_project(project.into_inner())
-        .await?;
-    Ok(web::Json(id))
-}
-
 /// Updates a project.
 /// This will create a new version.
 #[utoipa::path(
@@ -210,6 +161,173 @@ pub(crate) async fn delete_project_handler<C: ApplicationContext>(
     Ok(HttpResponse::Ok())
 }
 
+/// Retrieves details about the given version of a project.
+#[utoipa::path(
+    tag = "Projects",
+    get,
+    path = "/project/{project}/{version}",
+    responses(
+        (status = 200, description = "Project loaded from database", body = Project,
+            example = json!({
+                "id": "df4ad02e-0d61-4e29-90eb-dc1259c1f5b9",
+                "version": {
+                    "id": "8f4b8683-f92c-4129-a16f-818aeeee484e",
+                    "changed": "2021-04-26T14:05:39.677390600Z",
+                    "author": "5b4466d2-8bab-4ed8-a182-722af3c80958"
+                },
+                "name": "Test",
+                "description": "Foo",
+                "layers": [],
+                "plots": [],
+                "bounds": {
+                    "spatialReference": "EPSG:4326",
+                    "boundingBox": {
+                        "lowerLeftCoordinate": {
+                            "x": 0.0,
+                            "y": 0.0
+                        },
+                        "upperRightCoordinate": {
+                            "x": 1.0,
+                            "y": 1.0
+                        }
+                    },
+                    "timeInterval": {
+                        "start": 0,
+                        "end": 1
+                    }
+                },
+                "timeStep": {
+                    "granularity": "months",
+                    "step": 1
+                }
+            })
+        )
+    ),
+    params(
+        ("project" = ProjectId, description = "Project id"),
+        ("version" = ProjectVersionId, description = "Version id")
+    ),
+    security(
+        ("session_token" = [])
+    )
+)]
+pub(crate) async fn load_project_version_handler<C: ApplicationContext>(
+    project: web::Path<(ProjectId, ProjectVersionId)>,
+    session: C::Session,
+    app_ctx: web::Data<C>,
+) -> Result<impl Responder> {
+    let project = project.into_inner();
+    let id = app_ctx
+        .session_context(session)
+        .db()
+        .load_project_version(project.0, LoadVersion::Version(project.1))
+        .await?;
+    Ok(web::Json(id))
+}
+
+/// Retrieves details about the latest version of a project.
+#[utoipa::path(
+    tag = "Projects",
+    get,
+    path = "/project/{project}",
+    responses(
+        (status = 200, description = "Project loaded from database", body = Project,
+            example = json!({
+                "id": "df4ad02e-0d61-4e29-90eb-dc1259c1f5b9",
+                "version": {
+                    "id": "8f4b8683-f92c-4129-a16f-818aeeee484e",
+                    "changed": "2021-04-26T14:05:39.677390600Z",
+                    "author": "5b4466d2-8bab-4ed8-a182-722af3c80958"
+                },
+                "name": "Test",
+                "description": "Foo",
+                "layers": [],
+                "plots": [],
+                "bounds": {
+                    "spatialReference": "EPSG:4326",
+                    "boundingBox": {
+                        "lowerLeftCoordinate": {
+                            "x": 0.0,
+                            "y": 0.0
+                        },
+                        "upperRightCoordinate": {
+                            "x": 1.0,
+                            "y": 1.0
+                        }
+                    },
+                    "timeInterval": {
+                        "start": 0,
+                        "end": 1
+                    }
+                },
+                "timeStep": {
+                    "granularity": "months",
+                    "step": 1
+                }
+            })
+        )
+    ),
+    params(
+        ("project" = ProjectId, description = "Project id")
+    ),
+    security(
+        ("session_token" = [])
+    )
+)]
+pub(crate) async fn load_project_latest_handler<C: ApplicationContext>(
+    project: web::Path<ProjectId>,
+    session: C::Session,
+    app_ctx: web::Data<C>,
+) -> Result<impl Responder> {
+    let id = app_ctx
+        .session_context(session)
+        .db()
+        .load_project_version(project.into_inner(), LoadVersion::Latest)
+        .await?;
+    Ok(web::Json(id))
+}
+
+/// Lists all available versions of a project.
+#[utoipa::path(
+    tag = "Projects",
+    get,
+    path = "/project/{project}/versions",
+    responses(
+        (status = 200, description = "OK", body = [ProjectVersion],
+            example = json!([
+                {
+                    "id": "8f4b8683-f92c-4129-a16f-818aeeee484e",
+                    "changed": "2021-04-26T14:05:39.677390600Z",
+                    "author": "5b4466d2-8bab-4ed8-a182-722af3c80958"
+                },
+                {
+                    "id": "ced041c7-4b1d-4d13-b076-94596be6a36a",
+                    "changed": "2021-04-26T14:13:10.901912700Z",
+                    "author": "5b4466d2-8bab-4ed8-a182-722af3c80958"
+                }
+            ])
+        )
+    ),
+    params(
+        ("project" = ProjectId, description = "Project id")
+    ),
+    security(
+        ("session_token" = [])
+    )
+)]
+pub(crate) async fn project_versions_handler<C: ApplicationContext>(
+    session: C::Session,
+    app_ctx: web::Data<C>,
+    project: web::Path<ProjectId>,
+) -> Result<impl Responder> {
+    let versions = app_ctx
+        .session_context(session)
+        .db()
+        .list_project_versions(project.into_inner())
+        .await?;
+    Ok(web::Json(versions))
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -217,7 +335,8 @@ mod tests {
     use crate::contexts::{Session, SimpleApplicationContext, SimpleSession};
     use crate::handlers::ErrorResponse;
     use crate::util::tests::{
-        check_allowed_http_methods, create_project_helper, send_test_request, update_project_helper,
+        check_allowed_http_methods, create_project_helper, send_test_request,
+        update_project_helper, with_temp_context,
     };
     use crate::util::Identifier;
     use crate::workflows::workflow::WorkflowId;
@@ -915,6 +1034,92 @@ mod tests {
             "ProjectDeleteFailed",
             "Failed to delete the project.",
         )
+        .await;
+    }
+
+    #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
+    async fn load_version() {
+        with_temp_context(|app_ctx, _| async move {
+            let ctx = app_ctx.default_session_context().await.unwrap();
+
+            let session = ctx.session().clone();
+            let project = create_project_helper(&app_ctx).await;
+
+            let db = ctx.db();
+
+            db.update_project(update_project_helper(project))
+                .await
+                .unwrap();
+
+            let req = test::TestRequest::get()
+                .uri(&format!("/project/{project}"))
+                .append_header((header::CONTENT_LENGTH, 0))
+                .append_header((header::AUTHORIZATION, Bearer::new(session.id().to_string())));
+            let res = send_test_request(req, app_ctx.clone()).await;
+
+            assert_eq!(res.status(), 200);
+
+            let body: Project = test::read_body_json(res).await;
+
+            assert_eq!(body.name, "TestUpdate");
+
+            let versions = db.list_project_versions(project).await.unwrap();
+            let version_id = versions.first().unwrap().id;
+
+            let req = test::TestRequest::get()
+                .uri(&format!("/project/{project}/{version_id}"))
+                .append_header((header::CONTENT_LENGTH, 0))
+                .append_header((header::AUTHORIZATION, Bearer::new(session.id().to_string())));
+            let res = send_test_request(req, app_ctx).await;
+
+            assert_eq!(res.status(), 200);
+
+            let body: Project = test::read_body_json(res).await;
+            assert_eq!(body.name, "TestUpdate");
+        })
+        .await;
+    }
+
+    #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
+    async fn load_version_not_found() {
+        with_temp_context(|app_ctx, _| async move {
+            let session = app_ctx.default_session().await.unwrap();
+            let project = create_project_helper(&app_ctx).await;
+
+            let req = test::TestRequest::get()
+                .uri(&format!(
+                    "/project/{project}/00000000-0000-0000-0000-000000000000"
+                ))
+                .append_header((header::CONTENT_LENGTH, 0))
+                .append_header((header::AUTHORIZATION, Bearer::new(session.id().to_string())));
+            let res = send_test_request(req, app_ctx).await;
+
+            assert_eq!(res.status(), 400);
+        })
+        .await;
+    }
+
+    #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
+    async fn list_versions() {
+        with_temp_context(|app_ctx, _| async move {
+            let session = app_ctx.default_session().await.unwrap();
+
+            let project = create_project_helper(&app_ctx).await;
+
+            let ctx = app_ctx.session_context(session.clone());
+
+            let db = ctx.db();
+
+            db.update_project(update_project_helper(project))
+                .await
+                .unwrap();
+
+            let req = test::TestRequest::get()
+                .uri(&format!("/project/{project}/versions"))
+                .append_header((header::CONTENT_LENGTH, 0))
+                .append_header((header::AUTHORIZATION, Bearer::new(session.id().to_string())));
+            send_test_request(req, app_ctx).await;
+        })
         .await;
     }
 }
