+++ services/src/pro/apidoc.rs
@@ -123,12 +123,12 @@ use super::users::{UserCredentials, UserId, UserInfo, UserRegistration, UserSess
         handlers::datasets::suggest_meta_data_handler,
         handlers::spatial_references::get_spatial_reference_specification_handler,
         handlers::projects::list_projects_handler,
-        pro::handlers::projects::project_versions_handler,
+        handlers::projects::project_versions_handler,
+        handlers::projects::load_project_latest_handler,
         handlers::projects::create_project_handler,
-        pro::handlers::projects::load_project_latest_handler,
         handlers::projects::update_project_handler,
         handlers::projects::delete_project_handler,
-        pro::handlers::projects::load_project_version_handler,
+        handlers::projects::load_project_version_handler,
         handlers::upload::list_upload_files_handler,
         handlers::upload::list_upload_file_layers_handler,
         handlers::upload::upload_handler,
