+++ services/src/pro/handlers.rs
@@ -2,5 +2,4 @@ pub mod datasets;
 #[cfg(feature = "odm")]
 pub mod drone_mapping;
 pub mod permissions;
-pub mod projects;
 pub mod users;
