+++ services/src/pro/util/tests.rs
@@ -116,7 +116,7 @@ where
         .configure(handlers::layers::init_layer_routes::<C>)
         .configure(pro::handlers::permissions::init_permissions_routes::<C>)
         .configure(handlers::plots::init_plot_routes::<C>)
-        .configure(pro::handlers::projects::init_project_routes::<C>)
+        .configure(handlers::projects::init_project_routes::<C>)
         .configure(pro::handlers::users::init_user_routes::<C>)
         .configure(handlers::spatial_references::init_spatial_reference_routes::<C>)
         .configure(handlers::upload::init_upload_routes::<C>)
