+++ services/src/server.rs
@@ -1,9 +1,9 @@
 use crate::apidoc::ApiDoc;
-use crate::contexts::{PostgresContext, SimpleApplicationContext};
+use crate::contexts::{InMemoryContext, PostgresContext, SimpleApplicationContext};
 use crate::error::{Error, Result};
 use crate::handlers;
 use crate::util::config;
-use crate::util::config::get_config_element;
+use crate::util::config::{get_config_element, Backend};
 use crate::util::server::{
     calculate_max_blocking_threads_per_worker, configure_extractors, connection_init,
     log_server_info, render_404, render_405, serve_openapi_json, CustomRootSpanBuilder,
@@ -44,17 +44,26 @@ pub async fn start_server(static_files_dir: Option<PathBuf>) -> Result<()> {
 
     register_gdal_drivers_from_list(config::get_config_element::<config::Gdal>()?.allowed_drivers);
 
-    let ctx = match web_config.backend {
+    match web_config.backend {
         Backend::InMemory => {
             info!("Using in memory backend");
 
-            InMemoryContext::new_with_data(
+            let ctx = InMemoryContext::new_with_data(
                 data_path_config.dataset_defs_path,
                 data_path_config.provider_defs_path,
                 data_path_config.layer_defs_path,
                 data_path_config.layer_collection_defs_path,
                 tiling_spec,
                 chunk_byte_size,
+            )
+            .await;
+
+            start(
+                static_files_dir,
+                web_config.bind_address,
+                web_config.api_prefix,
+                web_config.version_api,
+                ctx,
             )
             .await
         }
@@ -71,7 +80,7 @@ pub async fn start_server(static_files_dir: Option<PathBuf>) -> Result<()> {
                 // fix schema by providing `search_path` option
                 .options(&format!("-c search_path={}", db_config.schema));
 
-            PostgresContext::new_with_data(
+            let ctx = PostgresContext::new_with_data(
                 pg_config,
                 tokio_postgres::NoTls,
                 data_path_config.dataset_defs_path,
@@ -81,18 +90,18 @@ pub async fn start_server(static_files_dir: Option<PathBuf>) -> Result<()> {
                 tiling_spec,
                 chunk_byte_size,
             )
-            .await?
-        }
-    };
+            .await?;
 
-    start(
-        static_files_dir,
-        web_config.bind_address,
-        web_config.api_prefix,
-        web_config.version_api,
-        ctx,
-    )
-    .await
+            start(
+                static_files_dir,
+                web_config.bind_address,
+                web_config.api_prefix,
+                web_config.version_api,
+                ctx,
+            )
+            .await
+        }
+    }
 }
 
 async fn start<C>(
